+++ test/test_config.py
@@ -2,7 +2,7 @@ from os import path
 from lambda_uploader import config
 
 EX_CONFIG = path.normpath(path.join(path.dirname(__file__),
-                          '../example'))
+                          '../test/configs'))
 
 
 def test_load_config():
@@ -22,6 +22,29 @@ def test_role_override():
     cfg = config.Config(EX_CONFIG, role=role)
 
     assert cfg.role is role
+
+
+def test_vpc_config():
+    subnet = 'subnet-00000000'
+    securitygroup = 'sg-00000000'
+    vpc = {
+        'subnets': [
+            subnet
+        ],
+        'security_groups': [
+            securitygroup
+        ]
+    }
+    cfg = config.Config(EX_CONFIG)
+
+    assert cfg.raw['vpc']['security_groups'][0] == vpc['security_groups'][0]
+    assert cfg.raw['vpc']['subnets'][0] == vpc['subnets'][0]
+
+
+def test_no_vpc():
+    cfg = config.Config(EX_CONFIG, EX_CONFIG + '/lambda-no-vpc.json')
+
+    assert cfg.raw['vpc'] is None
 
 
 def test_set_publish():